combine the watch manager test sets into a TestPulsar2 class
